fixed typecast error in hermite polynomials

+++ code/src/legendre.cpp
@@ -11,8 +11,8 @@ void Legendre::Compute() {
     Matrix CM( _degree, _degree, 0.0 );
 
     for( unsigned i = 0; i < _degree - 1; ++i ) {
-        CM( i + 1, i ) = std::sqrt( 1 / ( 4 - 1 / std::pow( i + 1, 2 ) ) );
-        CM( i, i + 1 ) = std::sqrt( 1 / ( 4 - 1 / std::pow( i + 1, 2 ) ) );
+        CM( i + 1, i ) = std::sqrt( 1 / ( 4 - 1 / std::pow( static_cast<double>( i + 1 ), 2 ) ) );
+        CM( i, i + 1 ) = std::sqrt( 1 / ( 4 - 1 / std::pow( static_cast<double>( i + 1 ), 2 ) ) );
     }
 
     auto evSys = MathTools::ComputeEigenValTriDiagMatrix( CM );
